move fetch user data from home screen to resource/committee screen due to bug
